+++ plugins/preact.ts
@@ -201,7 +201,7 @@ ${code}
 
     const assignBuchtaRoute = (code: string, route: string) => {
         return `
-const buchtaRoute = () => {
+globalThis.buchtaRoute = () => {
     let params = new Map();
 
     const path = "${route}";