+++ public/svelte/[id]/index.svelte
@@ -1,7 +1,10 @@
 <script>
     const route = buchtaRoute();
+    console.log(route)
+    import Main from "../components/Main.svelte";
 </script>
 
 <h1>
     I am at /svelte/{route.params.get("id")}/
 </h1>
+<Main />